rename handlers file to routes, relatesto issue #11 and PR #13
