+++ src/routes.js
@@ -1,6 +1,6 @@
-const { homeRoute } = require('./handlers/home.js');
-const { staticRoute } = require('./handlers/static.js');
-const { confirmationRoute } = require('./handlers/confirmation.js');
+const { homeRoute } = require('./routes/home.js');
+const { staticRoute } = require('./routes/static.js');
+const { confirmationRoute } = require('./routes/confirmation.js');
 
 const routes = [
   confirmationRoute,
