+++ src/routes/static.js
